#306 use int64 for config integer

+++ core/src/scheduler/optimizer/FaissFlatPass.h
@@ -45,9 +45,9 @@ class FaissFlatPass : public Pass {
     Run(const TaskPtr& task) override;
 
  private:
-    int32_t threshold_ = std::numeric_limits<int32_t>::max();
+    int64_t threshold_ = std::numeric_limits<int64_t>::max();
     int64_t count_ = 0;
-    std::vector<int32_t> gpus;
+    std::vector<int64_t> gpus;
 };
 
 using FaissFlatPassPtr = std::shared_ptr<FaissFlatPass>;
